Add mariusz app template
