--- examples/module1/lesson5/app-performance/loaders/clientSideApiLoader.ts
@@ -1,12 +1,0 @@
-import axios from 'axios';
-import { Bootstrap } from '../types/Bootstrap';
-
-export async function clientSideApiLoader(): Promise<Bootstrap> {
-  const { data } = await axios.get<Bootstrap>(
-    'http://localhost:3000/api/data/bootstrap?timeout=3000'
-  );
-
-  console.log('data: ', data);
-
-  return data;
-}
