--- examples/module1/lesson5/app-performance/loaders/serverSideApiLoader.ts
@@ -1,11 +1,0 @@
-import { Bootstrap } from '../types/Bootstrap';
-
-declare global {
-  interface Window {
-    OFE_BOOTSTRAP_DATA: Bootstrap;
-  }
-}
-
-export function serverSideApiLoader() {
-  return window.OFE_BOOTSTRAP_DATA;
-}
