+++ examples/module1/lesson5/app-performance/main.tsx
@@ -4,8 +4,10 @@ import { RouterProvider } from 'react-router-dom';
 import { router } from './router';
 
 const rootElement = document.getElementById('app')!;
-if (!rootElement.innerHTML) {
-  const root = ReactDOM.createRoot(rootElement);
+const root = ReactDOM.createRoot(rootElement);
 
-  root.render(<RouterProvider router={router} />);
-}
+root.render(
+  <React.StrictMode>
+    <RouterProvider router={router} />
+  </React.StrictMode>
+);
